Move the group by nothing test case to common
